Added function to convert data to geojson
@hykim3594 I was thinking about the discussion yesterday about the
matching of a data.js that contains a variable with a name to contains
the json formatted data to the call to the data in the leaflet map
function.

In this commit I just convert the sp::object to a character vector and
don''t write it to a file. Perhaps we can place the json directly in
the same page as the map thereby avoiding the added complexity of
dealing with calling file names and variable names in the other
file. Any disadvantage to using a "one file" approach rather than
multiple files?