Added a function to produce a point layer
This function produces a point layer to be used in a map. This is a
list with the name, title and the script stuff for the map. I figure
that we will need to pass the name and title onto other parts of the
map and therefore it is useful to have them available when passing the
layer on to the map.