Scripts functions and corrections
Added a scripts function that will be used to extract the scripts for
each layer.

Made correction to the generic names method. The argument to a generic.something
function must have the same name as its generic method; hence changed
to 'x'

Also added a test for  names and scripts