Added function to build the string to make the overlays
The overlays are the layers in the map. When you call the map function
on a sinlge layer you want that to be defined as the only layer in the
leaflet map. When you call on a list of layers you want all the layers
to be called as overlays in the map.