Added a function to generate getColor js
After discussions on friday we decided to design around always calling
a function to get the color of the points because thsi can be reused
to call the colour in the legend when we get that far.

Next step is to change the call in the map function and carry this
function in the map layer object.